# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Paradise Circus API",
    "version": "1.0.0",
    "description": "Workshop booking and community platform API for flow artists. Consumed by the web app and mobile app."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development"
    },
    {
      "url": "https://paradisecircus.com/api",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Clerk session token"
      }
    },
    "schemas": {
...
      "icsContentSchema": {
        "type": "string",
        "description": "ICS calendar file content (text/calendar)"
      },
...
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                }
              }
            }
          }
        }
      },
...
      "404": {
        "description": "Not Found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Resource not found"
                }
              }
            }
          }
        }
      },
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/calendar/{token}": {
      "get": {
        "operationId": "get-calendar-{token}",
        "summary": "Get calendar file for participation (ICS)",
        "description": "Returns an ICS calendar file for a booking. Requires valid confirmation token.",
        "tags": [
          "Calendar"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/icsContentSchema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/event/{id}/calendar": {
      "get": {
        "operationId": "get-event-{id}-calendar",
        "summary": "Get calendar file for event (ICS)",
        "description": "Returns an ICS calendar file for a published event. No booking required.",
        "tags": [
          "Event"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/icsContentSchema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  }
}